requirements.txt
streamlit==1.37.0
pandas==2.0.3
markdown2==2.4.8
xhtml2pdf==0.2.16
matplotlib==3.8.3
seaborn==0.12.2
scikit-learn==1.3.0
langchain-openai==0.0.8
python-dotenv==1.0.1
